refactor: JWT 발급 관련 패키지 및 클래스 이름 변경
